Improve SP upload performance, removing 'pops', while reducing transfer rate
Instead of preparing separate packets in advance, the entire file is encrypted in one go, and split at the moment of bulk transfer in discrete packets. This fixed an alignment problem that gave an audible 'pop' every packet boundary. (this was at 47 seconds in the previous version) Also, packetsize is reduced to 2048, the exact framesize for SP. I'm not sure wheter or not these changes afffect LP2 or LP4 uploads, as I have no material to test it with.

netmdcli now takes a song title as an argument for 'send', instead of using the filename